[refactor/Mode1Up] Copy Mode1Up methods out of BookReader
